back_inserter part of iterator header
I was trying to compile without including the <iterator> header using the visual studio community 2019 64 bit compiler, but when compiling, I got the errors:
error C2039: 'back_inserter': is not a member of 'std' 
message : see declaration of 'std'
error C3861: 'back_inserter': identifier not found 

I found that std::back_inserter is included in the <iterator> header, which was not included. 
https://www.cplusplus.com/reference/iterator/back_inserter/

The solution compiles fine when I add this.